use double

+++ src/main/kotlin/eu/aagsolutions/trip/service/geo/services/GeoPointService.kt
@@ -1,13 +1,12 @@
 package eu.aagsolutions.trip.service.geo.services
 
 import com.google.maps.GeoApiContext
-import eu.aagsolutions.trip.service.geo.model.GeoPoint
 import com.google.maps.GeocodingApi
-import java.math.BigDecimal
 import com.google.maps.errors.ApiException
-import java.io.IOException
 import eu.aagsolutions.trip.service.exceptions.BadGatewayException
+import eu.aagsolutions.trip.service.geo.model.GeoPoint
 import org.springframework.stereotype.Service
+import java.io.IOException
 
 
 /**
@@ -19,8 +18,8 @@ class GeoPointService(val context: GeoApiContext?) {
     fun findGeoPointForAddress(address: String): GeoPoint {
         try {
             val results = GeocodingApi.geocode(context, address).await()
-            return GeoPoint(address, BigDecimal.valueOf(results[0].geometry.location.lat)
-                    ,BigDecimal.valueOf(results[0].geometry.location.lng))
+            return GeoPoint(address, results[0].geometry.location.lat,
+                    results[0].geometry.location.lng)
         } catch (e: ApiException) {
             throw BadGatewayException("Google maps service access error", e)
         } catch (e: InterruptedException) {
